finished tests

+++ PS6_DAL/src/main/java/base/PersonDAL.java
@@ -19,7 +19,7 @@ public class PersonDAL {
 	public static PersonDomainModel addPerson(PersonDomainModel per) {
 		Session session = HibernateUtil.getSessionFactory().openSession();
 		Transaction tx = null;
-		int employeeID = 0;
+
 		try {
 			tx = session.beginTransaction();
 			session.save(per);
@@ -32,24 +32,24 @@ public class PersonDAL {
 			session.close();
 		}
 		return per;
-	}
+	} 
 
 	public static ArrayList<PersonDomainModel> getPersons() {
 		Session session = HibernateUtil.getSessionFactory().openSession();
 		Transaction tx = null;
-		StudentDomainModel perGet = null;		
-		ArrayList<PersonDomainModel> perp = new ArrayList<PersonDomainModel>();
-		
+		StudentDomainModel perGet = null;
+		ArrayList<PersonDomainModel> pers = new ArrayList<PersonDomainModel>();
+
 		try {
-			tx = session.beginTransaction();	
-			
-			List students = session.createQuery("FROM StudentDomainModel").list();
-			for (Iterator iterator = students.iterator(); iterator.hasNext();) {
+			tx = session.beginTransaction();
+
+			List persons = session.createQuery("FROM PersonDomainModel").list();
+			for (Iterator iterator = persons.iterator(); iterator.hasNext();) {
 				PersonDomainModel per = (PersonDomainModel) iterator.next();
-				perp.add(per);
+				pers.add(per);
 
 			}
-			
+
 			tx.commit();
 		} catch (HibernateException e) {
 			if (tx != null)
@@ -58,24 +58,27 @@ public class PersonDAL {
 		} finally {
 			session.close();
 		}
-		return perp;		
+		return pers;
+
 	}
 
 	public static PersonDomainModel getPerson(UUID perID) {
-				
 		Session session = HibernateUtil.getSessionFactory().openSession();
 		Transaction tx = null;
-		PersonDomainModel perGet = null;		
-		
+		PersonDomainModel perGet = null;
+
 		try {
-			tx = session.beginTransaction();	
-									
-			Query query = session.createQuery("from StudentDomainModel where studentId = :id ");
-			query.setParameter("id", perID.toString());
-			
+			tx = session.beginTransaction();
+
+			Query query = session.createQuery("from PersonDomainModel where PersonId = :id ");
+			query.setParameter("id", perID);
+
 			List<?> list = query.list();
-			perGet = (PersonDomainModel)list.get(0);
-			
+			if ((list != null) && (list.size() == 1)) {
+				perGet = (PersonDomainModel) list.get(0);
+			} else
+				return null;
+
 			tx.commit();
 		} catch (HibernateException e) {
 			if (tx != null)
@@ -90,15 +93,15 @@ public class PersonDAL {
 	public static void deletePerson(UUID perID) {
 		Session session = HibernateUtil.getSessionFactory().openSession();
 		Transaction tx = null;
-		PersonDomainModel perGet = null;		
-		
+		PersonDomainModel perGet = null;
+
 		try {
-			tx = session.beginTransaction();	
-									
+			tx = session.beginTransaction();
+
 			PersonDomainModel per = (PersonDomainModel) session.get(PersonDomainModel.class, perID);
-			session.delete(per);
-		
-			
+			if (per != null)
+				session.delete(per);
+
 			tx.commit();
 		} catch (HibernateException e) {
 			if (tx != null)
@@ -107,20 +110,18 @@ public class PersonDAL {
 		} finally {
 			session.close();
 		}
-
 	}
 
 	public static PersonDomainModel updatePerson(PersonDomainModel per) {
 		Session session = HibernateUtil.getSessionFactory().openSession();
 		Transaction tx = null;
-		PersonDomainModel perGet = null;		
-		
+		PersonDomainModel perGet = null;
+
 		try {
-			tx = session.beginTransaction();	
-									
+			tx = session.beginTransaction();
+
 			session.update(per);
-	
-			
+
 			tx.commit();
 		} catch (HibernateException e) {
 			if (tx != null)
